Retry entire attribution bug test for now.
It seems like there is still a race condition: we sometimes still miss
detecting localStorage tracking because localStorage is written to after
we check it.

It makes sense that this seems to happen only in Chrome; Chrome is
faster, it can inject and run our code fast enough for the condition
to happen, even with "document_idle" injection.

It's not clear to me why this seems to have started (gotten worse?)
after https://github.com/EFForg/privacybadger/pull/1742.

Perhaps increasing extension storage writing buffer to two seconds
(from 300 ms) acts to speed up the extension enough to surface this race
condition?