Fix slider handler margins
Need to use "this" inside the create callback because the slider child
of the switch toggle doesn't exist yet.
We add the margin to have the slider handle positioned at the beginning
of the red/yellow/green slider element rather than its center.
Since this is positioned inside the wider slider handle, the absolute position
needs to be calculated according to the blocked/cookieblocked/allowed value in JS.
It might be possible to avoid this 1) if the handler can work with its handler in
the center or 2) in CSS if the slider handle was made in a different way.

The this in the stop callback is also different, but it doesn't matter
because it is used to find a parent of both the switch toggle and slider handler.

Added comments popup.js too to avoid future confusion.