Add option to allow learning in incognito windows
If the user chooses to, treat all windows as if they are regular ones and keep
records on them for heuristics.
This is disabled by default because it results in the storage of user history
during a private session.

Renamed tabIsIncognito to dontLearnInTab to better reflect the value it returns
(false for incognito windows if the user has allowed learning in them).