Fix options slider test
Now that breakage warnings look only at the yellowlist,
we don't get a breakage warning in this slider test.

In the test, the test domain gets initially set to "cookieblock"
in action map w/o the domain (or any of its parents) being present
on the yellowlist.

This would not happen in actual use; domains only get cookieblocked
when they or their parent(s) are present on the yellowlist.

Following up on 07f9b8061b92bdd5fc410371396a63be7ade7a3f and
https://github.com/EFForg/privacybadger/pull/2545