Set the welcome page workaround flag right away
On first run, without waiting for managed storage.

This updates the disappearing welcome page workaround to account
for the welcome page now being delayed by a few seconds.

This will reopen the welcome page when the extension gets restarted
before the welcome page had a chance to be shown the first time.