Remove legacy Pinterest button replacement for now
Tends to insert a giant SVG into page head replacing the Pinterest
script tag. We shouldn't provide a simple type 0 placeholder
when it's just a script tag in the head.

It's also not clear we need a Pinterest button replacement at all
at this point. Depending on user feedback, we may want to restore
with an improved version.